Move test logger to own file.
Because I keep looking for this feature and wondering where it is and why the heck I didn't write it yet.  'Oh OBVIOUSLY logging is a "misc" thing' ... said no one ever.
