+++ testutil/misc.go
@@ -2,10 +2,6 @@ package testutil
 
 import (
 	"bytes"
-	"io"
-
-	"github.com/inconshreveable/log15"
-	"github.com/smartystreets/goconvey/convey"
 )
 
 /*
@@ -31,24 +27,4 @@ func AdditionalDescription(addtnlDesc ...string) string {
 	buf.WriteString(addtnlDesc[n-1])
 	buf.WriteRune(')')
 	return buf.String()
-}
-
-var _ io.Writer = Writer{}
-
-/*
-	Wraps a goconvey context into an `io.Writer` so that you can
-	shovel logs at it.
-*/
-type Writer struct {
-	Convey convey.C
-}
-
-func (lw Writer) Write(msg []byte) (int, error) {
-	return lw.Convey.Print(string(msg))
-}
-
-func TestLogger(c convey.C) log15.Logger {
-	log := log15.New()
-	log.SetHandler(log15.StreamHandler(Writer{c}, log15.TerminalFormat()))
-	return log
 }
